Added Chart Service registration.  Added multi-line Chart example visible in modal dialog in a Blazor Server app.

+++ AnthroCloud/AnthroCloud.UI.Blazor/Services/IChartService.cs
@@ -6,6 +6,6 @@ namespace AnthroCloud.UI.Blazor.Services
 {
     public interface IChartService
     {
-        Task<List<WeightForLength>> GetAllWFL(byte id, decimal x, decimal y);
+        Task<List<WeightForLength>> GetAllWFL(byte id, double x, double y);
     }
 }
